pyproject.toml
[project]
name = "power-indicator-light"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bleak>=2.1.1",
    "colour-science>=0.4.7",
    "dirigera>=1.2.6",
    "flask>=3.1.3",
    "pydantic-settings>=2.13.1",
]

[project.scripts]
start-app = "power_indicator_light.main:start"

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
